Sprite position is no longer in Sprite types ownership
Ownership of position is moved to the Sprite Animator. An obvious failure to spot the issue from my end.

Animations now control their own positioning in the world.

destRect is only accessible as a getter. Setter has been set to private as the user has no reason to be modifying the internal destRect.